requirements.txt
absl-py==0.9.0
affine==2.3.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
contextily==1.0.0
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Fiona==1.8.13.post1
gast==0.3.3
geographiclib==1.50
geopandas==0.8.0
geopy==2.0.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.14.1
grpcio==1.27.2
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.1
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.1
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.5
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-slack==2.1.4
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-tcn==3.1.1
kiwisolver==1.2.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.2
mercantile==1.1.5
metar==1.7.0
mistune==0.8.4
#mkl-fft==1.1.0
#mkl-random==1.1.1
#mkl-service==2.3.0
munch==2.5.0
#nb-conda-kernels==2.2.3
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip==20.1.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
#protobuf==3.12.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pydot==1.4.1
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyproj==2.6.1.post1
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
rasterio==1.1.5
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.0
ruptures==1.0.6
scikit-learn==0.23.1
scipy
seaborn==0.10.1
Send2Trash==1.5.0
#setuptools==47.3.1.post20200622
Shapely==1.7.0
six==1.15.0
snuggs==1.4.7
tabulate==0.8.9
tensorboard==2.2.1
#tensorboard-plugin-wit==1.6.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.34.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==1.2.0
yapf==0.30.0
zipp==3.1.0
